Update square root.py

+++ square root.py
@@ -2,7 +2,7 @@ print("Type a number and this application will find the square root.")
 print("This application was made by Sameer Achhab.")
 import math
 while True:
- num=int(input())
+ num=float(input())
  answer=math.sqrt(num)
  print(answer)
